// src/types/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Date: { input: unknown; output: unknown };
  DateTime: { input: unknown; output: unknown };
  Duration: { input: unknown; output: unknown };
  LocalDateTime: { input: unknown; output: unknown };
};
// ...
export type Event = {
  __typename: 'Event';
  /** Event creating datetime */
  createdTime: Maybe<Scalars['DateTime']['output']>;
  /** The scheduled time of the event */
  dateTime: Maybe<Scalars['DateTime']['output']>;
  /** Description of the event. Visibility may be limited for private groups */
  description: Scalars['String']['output'];
  /** Duration of the event */
  duration: Maybe<Scalars['Duration']['output']>;
  /** Time when event ends */
  endTime: Maybe<Scalars['DateTime']['output']>;
  /** List of event hosts */
  eventHosts: Maybe<Array<Maybe<EventHost>>>;
  /** Type of the event */
  eventType: EventType;
  /** Name used for the event's web address on meetup.com. Must be between 6 and 60 characters */
  eventUrl: Scalars['String']['output'];
  featuredEventPhoto: Maybe<PhotoInfo>;
  /** Optional organizer-defined fee payment settings */
  feeSettings: Maybe<EventFeeSettings>;
  /** the group this event belongs to */
  group: Maybe<Group>;
  guestLimit: Maybe<Scalars['Int']['output']>;
  guestsAllowed: Scalars['Boolean']['output'];
  /** Returns a connection for the RSVP's for the hosts of the event. */
  hostRsvps: HostRsvpConnection;
  /**
   * String describing the location event venue. Visibility may be limited for private groups
   * For online events this field is used for the event's url.
   */
  howToFindUs: Maybe<Scalars['String']['output']>;
  /**
   * A unique but unstable identifier for the event.
   * Recurring events that have see activity may change the value of this identifier.
   */
  id: Scalars['ID']['output'];
  /** True the current member attending this event */
  isAttending: Scalars['Boolean']['output'];
  /** True if the current member has saved this event */
  isSaved: Scalars['Boolean']['output'];
  /** Maximum amount of RSVP spots for the event */
  maxTickets: Maybe<Scalars['Int']['output']>;
  /** Network event details */
  networkEvent: Maybe<NetworkEvent>;
  numberOfAllowedGuests: Scalars['Int']['output'];
  photoAlbum: Maybe<EventPhotoAlbum>;
  /** Get the current member's RSVP for the event if any */
  rsvp: Maybe<Rsvp>;
  /** Get a list of RSVP questions for the event. Visibility may be limited */
  rsvpQuestions: Array<RsvpQuestion>;
  /** Search for members RSVP's in this event. Visibility may be limited for private groups */
  rsvpSearch: RsvpSearchConnection;
  rsvpSettings: RsvpOpenSettings;
  /**
   * Get the current member's RSVP state for the event
   * Uses own memberId if nothing was provided.
   */
  rsvpState: RsvpState;
  /** Connections of filterable rsvps. Full list of information may be limited for private groups */
  rsvps: RsvpConnection;
  /** If this event is part of a series, this field describes that series */
  series: Maybe<Series>;
  /** Speaker details */
  speakerDetails: Maybe<SpeakerDetails>;
  /** Current status of the event */
  status: EventStatus;
  /** The event title for  display */
  title: Scalars['String']['output'];
  /** A unique and stable identifier for the event. */
  token: Scalars['ID']['output'];
  /** A connection of topics for the event */
  topics: TopicsConnection;
  /**
   * Returns all venues for an event.
   * Will return a list of 1 venue for online or in person
   * Hybrid events will have 2 entries
   */
  venues: Maybe<Array<Venue>>;
  /** Waitlist handling when RSVP limit is reached */
  waitlistMode: WaitlistMode;
  /** A Zoom meeting ID if the event is online and uses Zoom integration. */
  zoomMeetingId: Maybe<Scalars['ID']['output']>;
};
// ...
export type EventFeeSettings = {
  __typename: 'EventFeeSettings';
  /** Acceptable methods of payment */
  accepts: PaymentMethod;
  /** Amount of the fee */
  amount: Scalars['Float']['output'];
  /** Currency accepted for fee */
  currency: EventFeeCurrency;
  /** Optional discounted price for early RSVPers. */
  earlyBirdDiscount: Maybe<EarlyBirdDiscount>;
  /** Return true if paid event and has active promo codes, false otherwise */
  hasPromoCodes: Scalars['Boolean']['output'];
  /**
   * Organizer-defined terms for refunds. If this is defined, you must provide the authenticated member a way to access this information before they can RSVP.
   * They will need to agree to these terms before they RSVP
   * See the following help article for more information on Meetup refund payment policies https://help.meetup.com/hc/en-us/articles/360004065471-Payment-Policies
   */
  refundPolicy: Maybe<RefundPolicy>;
  /** Indicates if a fee is required to RSVP */
  required: Scalars['Boolean']['output'];
};

export type EventFundraisingInput = {
  enabled: Scalars['Boolean']['input'];
};

export type EventHost = {
  __typename: 'EventHost';
  member: Maybe<Member>;
  /** @deprecated Use member instead */
  memberId: Maybe<Scalars['ID']['output']>;
  /** Group member photo */
  memberPhoto: Maybe<PhotoInfo>;
  name: Maybe<Scalars['String']['output']>;
};

export type EventPhotoAlbum = {
  __typename: 'EventPhotoAlbum';
  id: Scalars['ID']['output'];
  photoCount: Scalars['Int']['output'];
  title: Scalars['String']['output'];
};
// ...
export enum EventType {
  /** The event takes place both online and in-person */
  Hybrid = 'HYBRID',
  /** The event takes place online */
  Online = 'ONLINE',
  /** The event takes place in-person */
  Physical = 'PHYSICAL',
}
// ...
export type Group = {
  __typename: 'Group';
  /** A list of group active topics. */
  activeTopics: Array<Topic>;
  /** Indicates if members are allowed to upload photos. */
  allowMemberPhotoUploads: Scalars['Boolean']['output'];
  /** Shows current user's permission to upload photos to this group. */
  canAddPhotos: Scalars['Boolean']['output'];
  /** A city where a group is located. */
  city: Maybe<Scalars['String']['output']>;
  /** A country where a group is located. */
  country: Maybe<Scalars['String']['output']>;
  /** What this group calls its members */
  customMemberLabel: Maybe<Scalars['String']['output']>;
  /** Description of a group. */
  description: Maybe<Scalars['String']['output']>;
  /** Active group membership dues settings. Available only when dues are enabled in the group. */
  duesSettings: Maybe<DuesSettings>;
  /** an email address, representing an Organizer(s) of this Meetup Group. Meant to use for Organizer Announcements */
  emailAnnounceAddress: Maybe<Scalars['String']['output']>;
  /**
   * Search for events in this group
   * @deprecated use 'events' instead
   */
  eventSearch: GroupEventSearchConnection;
  /** Events of the group */
  events: GroupEventConnection;
  /** Returns a list of recently used event photos */
  featuredEventPhotos: FeaturedEventPhotoConnection;
  /** Date and time of when a group was created. */
  foundedDate: Maybe<Scalars['DateTime']['output']>;
  /** Groups analytics, only available for groups which are part of a pro network */
  groupAnalytics: Maybe<GroupAnalytics>;
  /** Group unique identifier. */
  id: Scalars['ID']['output'];
  /** Checks whether the specified member is in this group. */
  isMember: Scalars['Boolean']['output'];
  /** Checks whether the current member is the primary organizer of this group */
  isPrimaryOrganizer: Scalars['Boolean']['output'];
  /** If the group is private */
  isPrivate: Scalars['Boolean']['output'];
  /** Join mode for new group members. */
  joinMode: GroupJoinMode;
  /** A group featured (cover) photo. */
  keyGroupPhoto: Maybe<PhotoInfo>;
  /** A latitude of a group location. */
  lat: Maybe<Scalars['Float']['output']>;
  /** Link for the group on meetup.com */
  link: Maybe<Scalars['String']['output']>;
  /** A longitude of a group location. */
  lon: Maybe<Scalars['Float']['output']>;
  /** The authenticated member's membership within this group if any */
  membershipMetadata: Maybe<Membership>;
  /** Members in this group */
  memberships: GroupMemberConnection;
  /** Name of a group. */
  name: Scalars['String']['output'];
  /** A flag indicating whether new members are required to upload a photo to their group profile in order to join this group. */
  needsPhoto: Maybe<Scalars['Boolean']['output']>;
  /** A flag indicating whether new members are required to answer organizer's questions in order to join this group. */
  needsQuestions: Maybe<Scalars['Boolean']['output']>;
  /** The primary organizer of the group. Null if the group is leaderless. */
  organizer: Maybe<Member>;
  /** A date when group joined pro network */
  proJoinDate: Maybe<Scalars['DateTime']['output']>;
  /** If a group is a part of a Pro network, the network properties. */
  proNetwork: Maybe<ProNetwork>;
  /** If a group requires new members to answer questions, a list of questions. */
  questions: Array<GroupQuestion>;
  /** A list of links to this group in social networks. */
  socialNetworks: Array<SocialNetwork>;
  /** List of the sponsors for the group */
  sponsors: ProGroupSponsorsConnection;
  /** A state, county or province where a group is located. */
  state: Maybe<Scalars['String']['output']>;
  /** Statistics about group members. */
  stats: Maybe<GroupStats>;
  /** Group status. */
  status: GroupStatus;
  /** Timezone identifier as used by the IANA timezone database, for example: "US/Eastern", "Asia/Calcutta" */
  timezone: Maybe<Scalars['String']['output']>;
  /** A group topic category. */
  topicCategory: Maybe<TopicCategory>;
  /** Group unique URL path. */
  urlname: Scalars['String']['output'];
  /** Recently used in-person venues */
  venues: GroupVenueConnection;
  /** Group featured video. */
  video: Maybe<GroupVideo>;
  /** A group welcome message. */
  welcomeBlurb: Maybe<Scalars['String']['output']>;
  /** A zip code of a group location. */
  zip: Maybe<Scalars['String']['output']>;
};
// ...
export type HostRsvpConnection = {
  __typename: 'HostRsvpConnection';
  edges: Array<RsvpEdge>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type NetworkEvent = {
  __typename: 'NetworkEvent';
  /** Time network event was created */
  createdAt: Maybe<Scalars['DateTime']['output']>;
  /** Start time of network event */
  eventTime: Maybe<Scalars['DateTime']['output']>;
  /** Number of groups participating in the network event */
  groupCount: Scalars['Int']['output'];
  /** Id of Network Event */
  id: Scalars['ID']['output'];
  /** If network event has been announced */
  isAnnounced: Maybe<Scalars['Boolean']['output']>;
  /** Number of the network event attendees */
  rsvpCount: Scalars['Int']['output'];
  /** Status of network event */
  status: Maybe<Scalars['String']['output']>;
  /** Timezone of network event */
  timezone: Maybe<Scalars['String']['output']>;
  /** Title of network event */
  title: Maybe<Scalars['String']['output']>;
};
// ...
export enum PaymentMethod {
  Cash = 'CASH',
  Paypal = 'PAYPAL',
}

/** Standard photo information */
export type PhotoInfo = {
  __typename: 'PhotoInfo';
  baseUrl: Maybe<Scalars['String']['output']>;
  highResUrl: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  standardUrl: Maybe<Scalars['String']['output']>;
  thumbUrl: Maybe<Scalars['String']['output']>;
};
// ...
export type RsvpOpenSettings = {
  __typename: 'RsvpOpenSettings';
  rsvpCloseTime: Maybe<Scalars['DateTime']['output']>;
  rsvpOpenTime: Maybe<Scalars['DateTime']['output']>;
  rsvpsClosed: Maybe<Scalars['Boolean']['output']>;
};

export type RsvpQuestion = {
  __typename: 'RsvpQuestion';
  questionId: Scalars['ID']['output'];
  text: Scalars['String']['output'];
};

export type RsvpSearchConnection = {
  __typename: 'RsvpSearchConnection';
  edges: Array<RsvpEdge>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};